refactor(grpo/data.py, grpo/gsm8k.py, rl/data.py, rl/gsm8k.py, llama4/_transform.py, phi4/_model_builders.py, phi4/_tokenizer.py): update ModelTokenizer import paths
Change the import paths for ModelTokenizer and related functions to reflect their new location in the torchtune.modules.transforms.tokenizers module. This improves organizational clarity and resolves deprecation warnings by removing the old import paths.

chore(tokenizers/__init__.py): remove deprecated tokenizers module

Delete the deprecated __init__.py file in the tokenizers module, which was maintained for backward compatibility. This file is no longer needed as the import paths have been updated, and its removal prepares the codebase for future releases.
